メディアのメタ情報をデータベースに登録する処理を追加
