+++ ocaz-sandbox/pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
   "beautifulsoup4~=4.12.2",
   "click~=8.1.3",
   "more_itertools~=9.1.0",
+  "opencv-python~=4.7.0.72",
   "pymongo~=4.3.3",
   "python-magic~=0.4.27",
   "requests~=2.30.0",
