+++ ocaz-sandbox/src/ocaz_sandbox/resolve_resolution.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 
-import contextlib
 import json
 import logging
 import os
@@ -8,27 +7,7 @@ import random
 from typing import Dict
 
 import click
-import cv2
 import pymongo
-
-
-@contextlib.contextmanager
-def open_video_capture(url: str) -> cv2.VideoCapture:
-    video_capture = cv2.VideoCapture(url)
-    assert video_capture.isOpened()
-    try:
-        yield video_capture
-    finally:
-        video_capture.release()
-
-
-def get_video_info(video_capture: cv2.VideoCapture) -> Dict:
-    return {
-        "width": int(video_capture.get(cv2.CAP_PROP_FRAME_WIDTH)),
-        "height": int(video_capture.get(cv2.CAP_PROP_FRAME_HEIGHT)),
-        "numberOfFrames": int(video_capture.get(cv2.CAP_PROP_FRAME_COUNT)),
-        "fps": video_capture.get(cv2.CAP_PROP_FPS),
-    }
 
 
 @click.command()
@@ -47,40 +26,8 @@ def main(mongodb_url, limit):
     mongo_col_object = mongo_db["object"]
 
     for object_id_record in object_id_records:
-        object_id = object_id_record["_id"]
-        logging.info(f"object_id = {object_id}")
 
-        object_record = mongo_col_object.find_one(
-            {"_id": object_id},
-            {"_id": True, "mimeType": True, "image": True, "video": True},
-        )
-        if "image" in object_record or "video" in object_record:
-            logging.info("the record already has image/video info.")
-            continue
 
-        url_record = mongo_col_url.find_one({"head10mbSha1": object_id, "available": True}, {"url": True})
-        url = url_record["url"]
-
-        logging.info(f"get {url}")
-        with open_video_capture(url) as video_capture:
-            video_info = get_video_info(video_capture)
-
-        if object_record["mimeType"].startswith("image/"):
-            del video_info["numberOfFrames"]
-            del video_info["fps"]
-            new_object_record = {"image": video_info}
-        elif object_record["mimeType"].startswith("video/"):
-            video_info["duration"] = video_info["numberOfFrames"] / video_info["fps"]
-            new_object_record = {"video": video_info}
-
-        logging.info(f"new_object_record = {json.dumps(new_object_record)}")
-        mongo_col_object.update_one(
-            {"_id": object_id},
-            {
-                "$set": new_object_record,
-            },
-            upsert=True,
-        )
 
     logging.info("done")
